Refactor to address review comments
