+++ src/components/shared/index.ts
@@ -1,5 +1,4 @@
 export * from "./chip";
-export * from "./drag-upload";
 export * from "./error-messages";
 export * from "./prompt";
 export * from "./sanitize";
